refactor: change name and add docs
