+++ dfaligner/dataset.py
@@ -7,7 +7,7 @@ from typing import List, Optional
 import numpy as np
 import pytorch_lightning as pl
 import torch
-from smts.DeepForcedAligner.dfa.config import DFAlignerConfig
+from smts.DeepForcedAligner.dfaligner.config import DFAlignerConfig
 from smts.text import TextProcessor
 from torch.nn.utils.rnn import pad_sequence
 from torch.utils.data import random_split
@@ -129,7 +129,7 @@ class AlignerDataset(Dataset):
                 basename,
                 speaker,
                 language,
-                f"spec-{self.sampling_rate}-{self.config.preprocessing.audio.spec_type.value}.npy",
+                f"spec-{self.sampling_rate}-{self.config.preprocessing.audio.spec_type}.npy",
             )
             .squeeze()
             .transpose(0, 1)
