+++ setup.cfg
@@ -1,5 +1,5 @@
 [isort]
-known_first_party = dfa
+known_first_party = dfaligner
 multi_line_output=3
 include_trailing_comma=True
 force_grid_wrap=0
